Extract Audio Processor to separate File
